requirements/tests.txt
docutils
model_mommy
django-ttag>=2.3
django-discover-runner
coverage
flake8
